:zap: change types from int to float

+++ host/gameplay_settings.py
@@ -11,16 +11,16 @@ class ForeignGameplaySettings(BaseModel):
 
 class BankGameplaySettings(BaseModel):
     maximum_tax_rate: float
-    starter_funds: float
+    starter_funds: int
 
 
 class InteriorGameplaySettings(BaseModel):
     population_per_infrastructure: int
     revenue_per_population: float
     cashback_modifier: float
-    starter_infrastructure: float
+    starter_infrastructure: int
     starter_land: float
-    starter_technology: float
+    starter_technology: int
 
 
 class GameplaySettingsModel(BaseModel):
